+++ labbench/__init__.py
@@ -29,7 +29,7 @@
 from .util import concurrently, sequentially, Call, stopwatch, retry, until_timeout, show_messages, sleep, console, _force_full_traceback
 _force_full_traceback(True)
 
-from ._backends import ShellBackend, DotNetDevice, LabviewSocketInterface, SerialLoggingDevice, TelnetDevice, VISADevice, Win32ComDevice
+from ._backends import ShellBackend, DotNetDevice, LabviewSocketInterface, SerialDevice, SerialLoggingDevice, TelnetDevice, VISADevice, Win32ComDevice
 from ._data import CSVLogger, HDFLogger, SQLiteLogger, read
 from ._device import Device, list_devices
 from ._host import Email